kodadocs/pyproject.toml
[project]
name = "kodadocs"
version = "2.1.0"
description = "Claude Code MCP tool that auto-generates help documentation for web applications"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{name = "KodaDocs", email = "hello@kodadocs.com"}]
keywords = ["documentation", "ai", "mcp", "claude-code", "vitepress", "help-center", "screenshots"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer",
    "rich",
    "pydantic>=2.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python",
    "tree-sitter-javascript",
    "tree-sitter-typescript",
    "anthropic",
    "pillow",
    "playwright",
    "fastmcp>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock", "ruff"]

[project.scripts]
kodadocs = "kodadocs.main:app"

[project.urls]
Homepage = "https://github.com/kodadocs/kodadocs"
Repository = "https://github.com/kodadocs/kodadocs"
Issues = "https://github.com/kodadocs/kodadocs/issues"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests requiring a running browser (deselect with '-m \"not integration\"')",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "pro-kit/",
  ".claude/",
  ".codedocs/",
  ".kodadocs/",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
